Refactor Repo
