--- browsertests/dont_run_this_yet.js
@@ -1,6 +1,0 @@
-fixture`Getting Started`
-    .page`http://devexpress.github.io/testcafe/example`;
-
-test('My first test', async (t) => {
-  // Test code
-});
