+++ components/core-hello/core-hello.html
@@ -1,4 +1,45 @@
-<p id="message">
-  <span id="hello">Hello World </span>
-  <slot name="name">Default Text</slot>
-</p>
+﻿<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Core-Hello Web Component</title>
+  <link rel="stylesheet" type="text/css" href="core-hello.css">
+</head>
+
+<body>
+  <core-hello rainbow lang="en">
+    <span slot="name">Mark</span>
+  </core-hello>
+
+  <core-hello lang="en">
+    <span slot="name">Mark</span>
+  </core-hello>
+
+  <core-hello lang="zh">
+    <span slot="name">马克</span>
+  </core-hello>
+
+  <core-hello lang="hi">
+    <span slot="name">Mark</span>
+  </core-hello>
+
+  <core-hello lang="hi" rainbow>
+    <span slot="name">Mark</span>
+  </core-hello>
+
+  <core-button rounded size="medium" color="primary">
+    Submit
+  </core-button>
+
+  <core-button shadow size="large">
+    Submit
+  </core-button>
+
+  <script src="components/core-hello/core-hello.js"></script>
+  <script src="components/core-button/core-button.js"></script>
+</body>
+
+</html>
