+++ package.json
@@ -5,7 +5,7 @@
   "main": "foo.js",
   "scripts": {
     "test": "npx karma start config/karma.conf.js",
-    "browsertest": "testcafe saucelabs:chrome,saucelabs:safari browsertests/*.js",
+    "browsertest": "testcafe saucelabs:chrome,saucelabs:safari tests/browser/*.test.js",
     "lint": "npx eslint './**/*.js' && npx htmllint *.html && npx stylelint components",
     "docs": "npx jsdoc -r . -c jsdoc.json -d docs",
     "build": "npx webpack --config config/webpack.prod.js",
