+++ tests/browser/dont_run_this_yet.js
@@ -1,0 +1,8 @@
+import { Selector } from 'testcafe';
+
+fixture `CoreHello Browser Tests`
+    .page`../../components/core-hello/core-hello.html`;
+
+test('My first test', async (t) => {
+  // Test code
+});
